requirements.txt
Django~=3.1.5
djangorestframework~=3.12.2
django-import-export>=2.5.0
django-phonenumber-field>=5.0.0
phonenumbers>=8.12.0